+++ src/pages/_app.tsx
@@ -8,10 +8,10 @@ import { useRouter } from "next/router";
 import React, { useEffect, useState } from "react";
 import { Toaster } from "react-hot-toast";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import RoleGuard from "@/lib/utils/RoleGuard";
 import PopUp, { CustomError } from "@/components/Common/PopUp";
 import ErrorBoundary from "@/components/ErrorBoundary";
 import GNB from "@/components/GNB/GNB";
-import RoleGuard from "@/components/RoleGuard";
 import Tab from "@/components/Tab";
 import "@/styles/globals.css";
